fix(pack-brain): anchor date-only event_counts bounds in the display timezone
A date-only since/until on brain.event_counts anchored to midnight UTC
(.and_utc()), putting the window's edges up to a day off the calendar
the caller was naming — the same defect ADR-169 removed from gtd's
date-only due. Both surfaces now share one rule: a bare YYYY-MM-DD
resolves to the earliest instant of that calendar date in the
configured display timezone; a date-only until still rolls to the NEXT
day's earliest instant so the half-open window covers the whole named
day; a date the zone's calendar skips entirely is rejected with an
error naming the zone.

The anchoring machinery (anchor_date_to_earliest_instant, the bisecting
gap resolver, and the all-zone sweep) moves verbatim from khive-pack-gtd
into khive-runtime::time_anchor — packs do not depend on packs, and the
rule interprets the runtime's own display_timezone config. gtd now
imports it; behavior there is unchanged (its full suite passes, and the
ignored 597-zone sweep passes in its new home: 4299 gaps, 818 folds, 7
whole-day skips, 0 violations).

Tests pin the semantics with the display timezone held fixed per test:
New York anchoring (04:00Z, four seeded events pinning all four window
edges), a UTC-config control reproducing the old midnight-UTC behavior,
a Havana spring-forward gap date anchoring at its 05:00Z boundary
within the resolver's documented one-second granularity, and a skipped
Samoa date rejected naming the zone.
