+++ crates/khive-pack-brain/Cargo.toml
@@ -34,6 +34,7 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 uuid = { workspace = true }
 chrono = { workspace = true }
+chrono-tz = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 [features]
