+++ crates/khive-runtime/src/lib.rs
@@ -37,6 +37,7 @@ pub mod retrieval;
 pub mod runtime;
 pub mod secret_gate;
 pub(crate) mod secret_gate_finalizer;
+pub mod time_anchor;
 pub use khive_storage::usage;
 pub mod validation;
